Fix build error in visual studio

+++ examples/common/src/vbomesh.cpp
@@ -1,8 +1,4 @@
 #include "vbomesh.h"
-//#include "glutils.h"
-
-//#define uint unsigned int
-
 #include <iostream>
 using std::cout;
 using std::cerr;
@@ -149,7 +145,7 @@ void VBOMesh::center( vector<vec3> & points ) {
     vec3 minPoint = points[0];
 
     // Find the AABB
-    for( uint i = 0; i < points.size(); ++i ) {
+    for( size_t i = 0; i < points.size(); ++i ) {
         vec3 & point = points[i];
         if( point.x > maxPoint.x ) maxPoint.x = point.x;
         if( point.y > maxPoint.y ) maxPoint.y = point.y;
@@ -165,7 +161,7 @@ void VBOMesh::center( vector<vec3> & points ) {
                         (maxPoint.z + minPoint.z) / 2.0f );
 
     // Translate center of the AABB to the origin
-    for( uint i = 0; i < points.size(); ++i ) {
+    for( size_t i = 0; i < points.size(); ++i ) {
         vec3 & point = points[i];
         point = point - center;
     }
@@ -176,11 +172,11 @@ void VBOMesh::generateAveragedNormals(
         vector<vec3> & normals,
         const vector<int> & faces )
 {
-    for( uint i = 0; i < points.size(); i++ ) {
+    for( size_t i = 0; i < points.size(); i++ ) {
         normals.push_back(vec3(0.0f));
     }
 
-    for( uint i = 0; i < faces.size(); i += 3) {
+    for( size_t i = 0; i < faces.size(); i += 3) {
         const vec3 & p1 = points[faces[i]];
         const vec3 & p2 = points[faces[i+1]];
         const vec3 & p3 = points[faces[i+2]];
@@ -194,7 +190,7 @@ void VBOMesh::generateAveragedNormals(
         normals[faces[i+2]] += n;
     }
 
-    for( uint i = 0; i < normals.size(); i++ ) {
+    for( size_t i = 0; i < normals.size(); i++ ) {
         normals[i] = glm::normalize(normals[i]);
     }
 }
@@ -209,14 +205,14 @@ void VBOMesh::generateTangents(
     vector<vec3> tan1Accum;
     vector<vec3> tan2Accum;
 
-    for( uint i = 0; i < points.size(); i++ ) {
+    for( size_t i = 0; i < points.size(); i++ ) {
         tan1Accum.push_back(vec3(0.0f));
         tan2Accum.push_back(vec3(0.0f));
         tangents.push_back(vec4(0.0f));
     }
 
     // Compute the tangent vector
-    for( uint i = 0; i < faces.size(); i += 3 )
+    for( size_t i = 0; i < faces.size(); i += 3 )
     {
         const vec3 &p1 = points[faces[i]];
         const vec3 &p2 = points[faces[i+1]];
@@ -245,7 +241,7 @@ void VBOMesh::generateTangents(
         tan2Accum[faces[i+2]] += tan2;
     }
 
-    for( uint i = 0; i < points.size(); ++i )
+    for( size_t i = 0; i < points.size(); ++i )
     {
         const vec3 &n = normals[i];
         vec3 &t1 = tan1Accum[i];
@@ -302,7 +298,7 @@ void VBOMesh::storeVBO( const vector<vec3> & points,
             tcIdx += 2;
         }
     }
-    for( unsigned int i = 0; i < elements.size(); ++i )
+    for( size_t i = 0; i < elements.size(); ++i )
     {
         el[i] = elements[i];
     }
